chore: cleanup file structure
