--- src/components/Layout/index.js
@@ -1,1 +1,0 @@
-export { default } from './Layout.component';
