+++ src/components/Private/Private.component.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
 
-import { useAuth } from '../../providers/Auth';
+import { useAuth } from 'providers/Auth';
 
 function Private({ children, ...rest }) {
   const { authenticated } = useAuth();
