+++ src/components/Video/VideoSideBar/VideoSideBar.component.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import VideoList from '../VideoList/VideoList.component';
-import { useVideoContext } from '../../../providers/video';
+import { VideoList } from 'components/Video';
+import { useVideoContext } from 'providers/video';
 
 const FlexNav = styled.nav`
   display: flex;
