+++ src/components/Video/index.js
@@ -1,0 +1,4 @@
+export { default as VideoCard } from './VideoCard/VideoCard.component';
+export { default as VideoList } from './VideoList/VideoList.component';
+export { default as VideoSidebar } from './VideoSideBar/VideoSideBar.component';
+export { default as Video } from './Video';
