+++ src/components/index.js
@@ -1,0 +1,5 @@
+export { default as App } from './App/App.component';
+export { default as Layout } from './Layout/Layout.component';
+export { default as Navigation } from './Navigation/Navigation.component';
+export { default as Private } from './Private/Private.component';
+export { default as SearchBar } from './Search/Search.component';
