+++ src/components/styled/index.js
@@ -1,1 +1,2 @@
-export { default } from './TextInput.component';
+export { default as TextInput } from './TextInput.component';
+export { default as ToggleButton } from './ToggleButton.component';
