+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './components/App';
+import { App } from 'components';
 import './global.css';
 
 ReactDOM.render(
