+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export { default as HomePage } from './Home/Home.page';
+export { default as LoginPage } from './Login/Login.page';
+export { default as VideoDetailsPage } from './VideoDetails/VideoDetails.page';
+export { default as NotFoundPage } from './NotFound/NotFound.page';
